+++ rust/numaflow-core/src/typ.rs
@@ -1,10 +1,15 @@
 //! Type configuration trait for Numaflow components.
 
-use crate::shared::create_components::create_rate_limiter;
+use crate::config::components::ratelimit::RateLimitConfig;
+use crate::error::Error;
 use crate::{Result, error};
+use numaflow_throttling::state::OptimisticValidityUpdateSecs;
 use numaflow_throttling::state::store::in_memory_store::InMemoryStore;
 use numaflow_throttling::state::store::redis_store::{RedisMode, RedisStore};
-use numaflow_throttling::{NoOpRateLimiter, RateLimit, RateLimiter, WithDistributedState};
+use numaflow_throttling::{
+    NoOpRateLimiter, RateLimit, RateLimiter, TokenCalcBounds, WithDistributedState,
+};
+use std::time::Duration;
 use tokio_util::sync::CancellationToken;
 
 pub trait NumaflowTypeConfig: Send + Sync + Clone + 'static {
@@ -61,7 +66,7 @@ pub async fn build_redis_rate_limiter(
 
 /// Build a Redis-backed rate limiter configuration.
 pub async fn build_redis_rate_limiter_config(
-    rate_limit_config: &crate::config::components::ratelimit::RateLimitConfig,
+    rate_limit_config: &RateLimitConfig,
     cln_token: CancellationToken,
 ) -> Result<WithRedisRateLimiter> {
     let limiter = build_redis_rate_limiter(rate_limit_config, cln_token).await?;
@@ -72,7 +77,7 @@ pub async fn build_redis_rate_limiter_config(
 
 /// Build an in-memory rate limiter from rate limit config
 pub async fn build_in_memory_rate_limiter(
-    rate_limit_config: &crate::config::components::ratelimit::RateLimitConfig,
+    rate_limit_config: &RateLimitConfig,
     cln_token: CancellationToken,
 ) -> Result<RateLimit<WithDistributedState<InMemoryStore>>> {
     // Create in-memory store
@@ -83,7 +88,7 @@ pub async fn build_in_memory_rate_limiter(
 
 /// Build an in-memory rate limiter configuration.
 pub async fn build_in_memory_rate_limiter_config(
-    rate_limit_config: &crate::config::components::ratelimit::RateLimitConfig,
+    rate_limit_config: &RateLimitConfig,
     cln_token: CancellationToken,
 ) -> Result<WithInMemoryRateLimiter> {
     let limiter = build_in_memory_rate_limiter(rate_limit_config, cln_token).await?;
@@ -100,12 +105,40 @@ pub async fn build_noop_rate_limiter_config() -> Result<WithoutRateLimiter> {
 }
 
 /// Helper function to determine which rate limiter to use based on the rate limit configuration.
-pub fn should_use_redis_rate_limiter(
-    rate_limit_config: &crate::config::components::ratelimit::RateLimitConfig,
-) -> bool {
+pub fn should_use_redis_rate_limiter(rate_limit_config: &RateLimitConfig) -> bool {
     rate_limit_config
         .store
         .as_ref()
         .and_then(|s| s.redis_store.as_ref())
         .is_some()
+}
+
+/// Creates rate limiter for a given store and rate limit configuration.
+pub async fn create_rate_limiter<S>(
+    rate_limit_config: &RateLimitConfig,
+    store: S,
+    cancel_token: CancellationToken,
+) -> error::Result<RateLimit<WithDistributedState<S>>>
+where
+    S: numaflow_throttling::state::Store + Sync + 'static,
+{
+    let bounds = TokenCalcBounds::new(
+        rate_limit_config.max,
+        rate_limit_config.min,
+        rate_limit_config.ramp_up_duration,
+    );
+
+    let refresh_interval = Duration::from_millis(100);
+    let runway_update = OptimisticValidityUpdateSecs::default();
+
+    RateLimit::<WithDistributedState<S>>::new(
+        bounds,
+        store,
+        rate_limit_config.processor_id,
+        cancel_token,
+        refresh_interval,
+        runway_update,
+    )
+    .await
+    .map_err(|e| Error::Config(format!("Failed to create rate limiter: {}", e)))
 }
